pyproject.toml
[project]
name = "worldfork"
version = "0.1.0"
description = "LLM-based interactive worldview game (Korean primary)"
authors = [{name = "hyunlord"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    # LLM (HARNESS_LAYER1 외부 패키지 정책 따름)
    "anthropic>=0.39.0",
    "httpx>=0.27.0",

    # 인프라
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "hypothesis>=6.0",
    "ruff>=0.1.0",
    "mypy>=1.7",
    "types-PyYAML>=6.0",
]
tier_1 = [
    # Tier 1 진입 시 추가 (지금은 X)
    "openai>=1.0",
    "google-generativeai>=0.3",
    "sqlalchemy>=2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=core --cov=service --cov=tools --cov-report=term-missing"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "service*", "tools*"]
